requirements.txt
hydra-core==1.3.2
matplotlib==3.5.1
numpy==1.22.3
omegaconf==2.3.0
Pillow==9.0.1
# Pillow==9.5.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.63.0
